requirements.txt
Flask
pyserial
schedule
